author note and level info in sidebar
